+++ nimare/annotate/ontology/cogat.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 
 from . import utils
-from ..text import uk_to_us
+from ...utils import uk_to_us
 from ...due import due
 from ... import references
 from ...extract import download_cognitive_atlas
