+++ nimare/tests/test_cogat.py
@@ -7,7 +7,7 @@ import pytest
 import pandas as pd
 
 import nimare
-from nimare import annotate
+from nimare import annotate, extract
 from .utils import get_test_data_path
 
 
@@ -18,7 +18,9 @@ def test_cogat():
     # A small test dataset with abstracts
     ns_dset_laird = nimare.dataset.Dataset.load(
         op.join(get_test_data_path(), 'neurosynth_laird_studies.pkl.gz'))
-    id_df, rel_df = annotate.ontology.cogat.download_cogat(overwrite=False)
+    cogat = extract.download_cognitive_atlas(overwrite=False)
+    id_df = pd.read_csv(cogat['ids'])
+    rel_df = pd.read_csv(cogat['relationships'])
     weights = {'isKindOf': 1, 'isPartOf': 1, 'inCategory': 1}
     counts_df, rep_text_df = annotate.ontology.cogat.extract_cogat(
         ns_dset_laird.texts, id_df, text_column='abstract')
